Remove advertising clause from BSD
We have a project that relies on this package, but the the 4-clause BSD license raises license compatibility and compliance issues. We asked the original developer of the OpenBSD Blowfish code for permission to remove the advertising clause, and he agreed (https://github.com/symphonyoss/hubot-symphony/issues/35#issuecomment-276809459). This change is to remove the clause upstream.